Search and replace to update datatype of all memory address trackers to unsigned long long to allow for 64 bit or larger machines.

+++ BusRequest.h
@@ -5,7 +5,7 @@ struct BusRequest {
     unsigned int commandCode;
     unsigned int targetAddress;
     unsigned int sourceAddress;
-    unsigned int payload;
+    unsigned long long payload;
 };
 
 #endif //BUSREQUEST_H
