+++ DummyNode.cpp
@@ -6,7 +6,7 @@ void DummyNode::acceptBusTransaction(BusRequest* d) {
     unsigned int commandCode = d->commandCode;
     unsigned int targetAdx = d->targetAddress;
     unsigned int sourceAdx = d->sourceAddress;
-    unsigned int payload   = d->payload;
+    unsigned long long payload   = d->payload;
     
     std::cout << " Got Request to my address: " << getAddress() << " From Source: " << sourceAdx <<" with code: " << commandCode << " and payload " << payload << std::endl;
 }
@@ -28,7 +28,7 @@ void DummyNode::setHasRequest(bool value) {
     hasRequest = value;
 }
     
-void DummyNode::sendNewTransaction(unsigned int adx, unsigned int code, unsigned int payload) {
+void DummyNode::sendNewTransaction(unsigned int adx, unsigned int code, unsigned long long payload) {
     BusRequest* r = new BusRequest;
     r->sourceAddress = getAddress();
     r->targetAddress = adx;
